WU: riportato all'origine il file
