+++ server.py
@@ -1,16 +1,26 @@
 import os
+from flask import Flask, request, jsonify
+from flask_cors import CORS
 import requests
 import aiohttp
 import urllib.parse
 import asyncio
-import json
 from dotenv import load_dotenv
 
 load_dotenv()
 
 ors_api_key = os.getenv("ORS_API_KEY")
 
-def getroutejson(startaddress, endaddress, routemode):
+app = Flask(__name__)
+
+# Enable CORS for all routes
+CORS(app, origins=["http://127.0.0.1:5000"])
+
+@app.route('/routejson')
+def routejson():
+    startaddress = request.args.get("startaddress")
+    endaddress = request.args.get("endaddress")
+    routemode = request.args.get("routemode")
 
     if (startaddress==None or endaddress==None):
         return {
@@ -25,7 +35,7 @@ def getroutejson(startaddress, endaddress, routemode):
     try:
         coords = asyncio.run(get_coordinates(startaddress, endaddress))
     except:
-        return json.dumps({
+        return jsonify({
             "error": "Non è stato possibile ricavare le coordinate geografiche dagli indirizzi forniti"
         })
 
@@ -47,14 +57,13 @@ def getroutejson(startaddress, endaddress, routemode):
 
     try:
         response = requests.post(openrouteservice_url, json=request_body, headers=headers)
-        response.raise_for_status()  
+        response.raise_for_status()
 
         data = response.json()
         return data
     except requests.exceptions.RequestException as error:
         print("Errore nella ricerca del percorso:", error)
-        print(f"API KEY: {ors_api_key}")
-        return json.dumps({
+        return jsonify({
             "error": f"Errore nella richiesta: {str(error)}",
         })
 
@@ -127,3 +136,6 @@ async def get_coordinates(address_start, address_end):
     }
 
     return coords
+
+if __name__ == '__main__':
+    app.run('127.0.0.1', 5001, debug=True)
